third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-video-element/video-loading-poster-deferred.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Video with loading=lazy that is not visible in viewport does not load poster image
  assert_equals: No network requests should be made for poster image expected 0 but got 1
Harness: the test ran to completion.
